setting and getting avatar refactoring
